_example/tsp: use floating points for coords and distances

+++ _example/tsp/utils.go
@@ -16,7 +16,7 @@ func max[T constraints.Ordered](a, b T) T {
 	return b
 }
 
-func worldBounds(cities []point) (maxw, maxh int) {
+func worldBounds(cities []point) (maxw, maxh float64) {
 	for _, c := range cities {
 		maxw = max(maxw, c.X)
 		maxh = max(maxw, c.Y)
